refactor(shared): split GuildAutomationService into Orchestrator + Repository
- Create IGuildAutomationRepository interface for DB operations
- Implement GuildAutomationRepository with all Prisma calls
- Extract business logic into GuildAutomationOrchestrator
- Slim service.ts to create dependency graph and export singleton

Public API unchanged: guildAutomationService methods work identically.
All tests pass. TypeScript strict mode enforced.
